lxc for wrlinux: Add local patches to 2.0.8 for LIN10
Issue: LIN10-3173

Add template script for WRLinux.
Add container name check in lxc-wrlinux.
Add defualt configuration file of lxc-net for future.
Tweak the script to make it work for both systemd and sysvinit.

(LOCAL REV: NOT UPSTREAM) -- not suitable for upstream: These patches are just
for generating WRLinux container for LXC. Some configuration is not suitable
for upstream.

Signed-off-by: Hongzhi.Song <hongzhi.song@windriver.com>