lxc: Add wrlinux templates to v2.0.8
Issue: LIN1018-348

Add template script for WRLinux.
Add container name check in lxc-wrlinux.
Tweak the script to make it work for both systemd and sysvinit.

(LOCAL REV: NOT UPSTREAM) -- not suitable for upstream: These patches are just
for generating WRLinux container for LXC. Some configuration is not suitable
for upstream.

Signed-off-by: Hongzhi.Song <hongzhi.song@windriver.com>